+++ examples/static_matrix.rs
@@ -19,4 +19,5 @@ fn main() {
     println!("a * b = {:?}", a.mul_matrix(&b));
     println!("a^T = {:?}", a.transpose());
     println!("det(a) = {:?}", a.determinant());
+    println!("rank(a) = {:?}", a.rank());
 }