+++ tests/matrix.rs
@@ -23,6 +23,7 @@ fn static_matrix_construction_and_operations() {
         StaticMatrix::new([[1.0, 4.0], [2.0, 5.0], [3.0, 6.0]])
     );
     assert_eq!(a.determinant(), -2.0);
+    assert_eq!(a.rank(), 2);
 }
 
 #[cfg(feature = "alloc")]
@@ -59,4 +60,5 @@ fn dynamic_matrix_construction_and_operations() {
         DynamicMatrix::new(3, 2, vec![1.0, 4.0, 2.0, 5.0, 3.0, 6.0]).unwrap()
     );
     assert_eq!(a.determinant(), Ok(-2.0));
+    assert_eq!(a.rank(), 2);
 }